Remove old config
